Add TransferRepository and related changes
