+++ src/ProjectOrigin.WalletSystem.Server/Models/TransferFilter.cs
@@ -5,6 +5,8 @@ namespace ProjectOrigin.WalletSystem.Server.Models;
 public record TransferFilter
 {
     public required string Owner { get; init; }
+    public int Skip { get; init; } = 0;
+    public int Limit { get; init; } = int.MaxValue;
     public DateTimeOffset? Start { get; init; }
     public DateTimeOffset? End { get; init; }
 }
