+++ src/ProjectOrigin.WalletSystem.Server/Repositories/ICertificateRepository.cs
@@ -20,11 +20,6 @@ public interface ICertificateRepository
     Task<IList<WalletSlice>> ReserveQuantity(string owner, string registryName, Guid certificateId, uint reserveQuantity);
     Task SetWalletSliceState(Guid sliceId, WalletSliceState state);
 
-    Task<IEnumerable<TransferViewModel>> GetTransfers(TransferFilter filter);
-    Task InsertTransferredSlice(TransferredSlice newSlice);
-    Task<TransferredSlice> GetTransferredSlice(Guid sliceId);
-    Task SetTransferredSliceState(Guid sliceId, TransferredSliceState state);
-
     Task InsertWalletAttribute(Guid walletId, WalletAttribute walletAttribute);
     Task<IEnumerable<WalletAttribute>> GetWalletAttributes(Guid walletId, Guid certificateId, string registryName, IEnumerable<string> keys);
 }
